tests: added checkpoint test improve utils
