+++ rl_coach/tests/utils/test_utils.py
@@ -18,7 +18,9 @@
 import glob
 import sys
 import time
+import os
 from os import path
+from rl_coach.tests.utils.definitions import Definitions as Def
 
 
 def print_progress(averaged_rewards, last_num_episodes, start_time, time_limit,
@@ -73,3 +75,65 @@ def read_csv_paths(test_path, filename_pattern, read_csv_tries=120,
             extra_tries -= 1
 
     return csv_paths
+
+
+def get_files_from_dir(dir_path):
+    """
+    Check if folder has files
+    :param dir_path: |string| folder path
+    :return: |list| return files in folder
+    """
+    start_time = time.time()
+    entities = None
+    while time.time() - start_time < Def.TimeOuts.wait_for_files:
+        # wait until logs created
+        if os.path.exists(dir_path):
+            entities = os.listdir(dir_path)
+            if len(entities) > 0:
+                break
+        time.sleep(1)
+
+    assert len(entities) > 0, \
+        Def.Consts.ASSERT_MSG.format("num files > 0", len(entities))
+    return entities
+
+
+def find_string_in_logs(log_path, str, timeout=Def.TimeOuts.wait_for_files):
+    """
+    Find string into the log file
+    :param log_path: |string| log path
+    :param str: |string| search text
+    :param timeout: |int| timeout for searching on file
+    :return: |bool| true if string found in the log file
+    """
+    start_time = time.time()
+    while time.time() - start_time < timeout:
+        # wait until logs created
+        if os.path.exists(log_path):
+            break
+        time.sleep(1)
+
+    if not os.path.exists(log_path):
+        return False
+
+    with open(log_path, 'r') as fr:
+        if str in fr.read():
+            return True
+    return False
+
+
+def get_csv_path(clres, tries_for_csv=Def.TimeOuts.wait_for_csv,
+                 extra_tries=0):
+    """
+    Get the csv path with the results - reading csv paths will take some time
+    :param clres: object of files that test is creating
+    :param tries_for_csv: timeout of tires until getting all csv files
+    :param extra_tries: add number of extra tries to check after getting all
+                        the paths.
+    :return: |list| csv path
+    """
+    return read_csv_paths(test_path=clres.exp_path,
+                          filename_pattern=clres.fn_pattern,
+                          read_csv_tries=tries_for_csv,
+                          extra_tries=extra_tries)
+
